add label text & classes, colors

+++ myLabeler.py
@@ -3,37 +3,54 @@ import numpy as np
 from glob import glob
 
 class Labeler:
-    def __init__(self, imagePath, labelPath):
-        self.imagePath = imagePath
-        self.labelPath = labelPath
-        self.imageList = self.getImageList()
-        self.fileIdx = 0
-        self.labelList = []
-        self.startPt = None
-        self.cpy = None
+    def __init__(self):
+        self.imagePath = 'images' # 이미지 디렉토리
+        self.labelPath = 'labels' # 라벨 디렉토리
+        self.imageList = self.getImageList() # 이미지 리스트
+        self.fileIdx = 0 # 현재 이미지 인덱스
+        self.labelList = [] # 라벨 리스트
+        self.startPt = None # 시작 좌표
+        self.cpy = None # 이미지 복사본
         self.colors = [(0, 0, 255),  # 빨강
                        (0, 165, 255),  # 주황
                        (0, 255, 255),  # 노랑
                        (0, 255, 0),  # 초록
                        (255, 0, 0),  # 파랑
-                       (255, 0, 255),  # 남색 
+                       (255, 0, 255),  # 남색
                        (128, 0, 128)  # 보라
                        ]
+        self.classes = ['chair', 'table', 'sofa', 'desk', 'monitor', 'planter', 'keyboard']
+        self.nowColor = 0 # 현재 색상 인덱스
     
-    # images에 jpg 파일 확인
-    # jpg 파일에 대응 되는 label txt 파일들 확인 -> 없으면 생성
+    # images 디렉토리에 있는 jpg 파일과 label 디렉토리에 있는 txt 파일들을 확인
     def getImageList(self):
         info = []
         
         # 현재 작업 디렉토리 확인
-        basePath = os.getcwd()
-        dataPath = os.path.join(basePath, self.imagePath)
-        imgNames = glob(os.path.join(dataPath, '*.jpg'))
+        basePath = os.getcwd() 
+        imagePath = os.path.join(basePath, self.imagePath)
+        labelPath = os.path.join(basePath, self.labelPath)
+
+        # images 디렉토리 확인
+        if not os.path.exists(imagePath):
+            # 종료
+            print("No images found in the directory")
+            sys.exit()
+
+        # label 디렉토리 생성
+        if not os.path.exists(labelPath):
+            os.makedirs(labelPath)
+            print("labels directory created")
         
+        # images 디렉토리에 있는 jpg 파일 확인
+        imgNames = glob(os.path.join(imagePath, '*.jpg'))
+        
+        # jpg 파일에 대응 되는 label txt 파일들 확인 -> 없으면 생성
         for img in imgNames:
             label = img.replace('images', 'labels')
             label = label.replace('.jpg', '.txt')
             
+            # label 파일이 없으면 생성
             if not os.path.exists(label):
                 with open(label, 'w') as f:
                     pass
@@ -56,13 +73,29 @@ class Labeler:
             # 경계 상자 그리기 (색상, 두께 등은 원하는대로 설정 가능)
             cv2.rectangle(self.cpy, (x1, y1), (x2, y2), line_color, thickness=lineWidth)
 
+            # 라벨 이름 표시
+            label_name = self.classes[class_id]
+            font = cv2.FONT_HERSHEY_SIMPLEX
+            font_scale = 0.71
+            font_thickness = 1
+            text_size, _ = cv2.getTextSize(label_name, font, font_scale, font_thickness)
+            text_x = x1
+            text_y = y1 - 10
+            
+            # 텍스트 테두리 사각형 그리기
+            box_coords = ((text_x, text_y - text_size[1] - 5), (text_x + text_size[0] + 5, text_y + 5))
+            cv2.rectangle(self.cpy, box_coords[0], box_coords[1], line_color, cv2.FILLED)
+
+            # 텍스트 그리기
+            cv2.putText(self.cpy, label_name, (text_x, text_y), font, font_scale, (255, 255, 255), font_thickness, cv2.LINE_AA)
+
     def onMouse(self, event, x, y, flags, param):
         if event == cv2.EVENT_LBUTTONDOWN:
             self.drawROI(param[0])
             self.startPt = (x, y)
-            self.labelList.append([0, self.startPt[0], self.startPt[1], 0, 0])
+            self.labelList.append([self.nowColor, self.startPt[0], self.startPt[1], 0, 0])
         elif event == cv2.EVENT_LBUTTONUP:
-            if self.labelList[-1][0] == x and self.labelList[-1][1] == y:
+            if self.labelList[-1][0] == x and self.labelList[-1][1] == y: 
                 self.labelList.pop()
                 self.drawROI(param[0])
             self.startPt = None
@@ -72,7 +105,7 @@ class Labeler:
                 self.labelList[-1][4] = y
                 self.drawROI(param[0])
         cv2.imshow('label', cv2.addWeighted(param[0], 0.3, self.cpy, 0.7, 0))
-    
+
     def convert_to_yolo_format(self, img, label):
         """두 점 좌표를 YOLO 형식으로 변환합니다.
 
@@ -187,8 +220,24 @@ class Labeler:
                     self.labelList.pop()
                     self.drawROI(img)
                     cv2.imshow('label', self.cpy)
+                elif key == ord('1'): # 빨강
+                    self.nowColor = 0
+                    print(self.nowColor)
+                elif key == ord('2'): # 주황
+                    self.nowColor = 1
+                    print(self.nowColor)
+                elif key == ord('3'): # 노랑
+                    self.nowColor = 2
+                elif key == ord('4'): # 초록
+                    self.nowColor = 3
+                elif key == ord('5'): # 파랑
+                    self.nowColor = 4
+                elif key == ord('6'): # 남색
+                    self.nowColor = 5
+                elif key == ord('7'): # 보라
+                    self.nowColor = 6
 
 if __name__ == '__main__':
-    l = Labeler('images', 'labels')
+    l = Labeler()
     
     l.run()
